companion_computer/requirements.txt
# Python Dependencies for Raspberry Pi Companion Computer

# Core
numpy>=1.21.0
PyYAML>=6.0

# Computer Vision
opencv-python>=4.5.0
picamera2>=0.3.0  # Pi Camera V2 interface

# AI/ML
tflite-runtime>=2.11.0  # TensorFlow Lite for edge inference
# Alternative: tensorflow>=2.11.0  # Full TF if needed

# Serial Communication
pyserial>=3.5
pymavlink>=2.4.0  # MAVLink protocol

# Networking
requests>=2.28.0

# Logging & Monitoring
loguru>=0.7.0

# GPS
gpsd-py3>=0.3.0  # If using GPSD daemon
pynmea2>=1.18.0  # NMEA parsing

# Utilities
python-dateutil>=2.8.0

# Scheduler (for mission scheduling)
schedule>=1.2.0

# Geofencing (for polygon operations)
shapely>=2.0.0

# Map visualization (optional for tools)
folium>=0.14.0
